ReadStream: handle multi-frame trailers correctly (+ move dataconn.Conn to pkg stream)
About channel becoming a paramter:
We use readStream recursively for reading the multi-frame trailer.
We also used buffered read channels.
Thus, as soon as readStream would return, the pending frames in the
channel would be lost, which was actually a major general error.

The former ReadStream is now private readStream, and we pass in the
channel as an argument.